requirements.txt
boto3==1.12.31
python-json-logger==0.1.11
PyYAML==5.3.1
pytest=6.1.2
